+++ index.d.ts
@@ -62,6 +62,13 @@ declare const THEME_ATTRIBUTE: CanUndef<string>;
  */
 declare const AVAILABLE_THEMES: CanUndef<string[]>;
 
+declare const DETECT_USER_PREFERENCES: CanUndef<
+	Dictionary<Dictionary<{
+		enabled: boolean;
+		aliases?: Dictionary<string>;
+	}>>
+>;
+
 interface Event {
 	delegateTarget?: Element;
 }