+++ src/super/i-static-page/modules/theme/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.60.0 (2023-09-28)
+
+#### :rocket: New Feature
+
+* Added possibility to detect the theme based on user system settings
+
 ## v3.0.0-rc.164 (2021-03-22)
 
 #### :rocket: New Feature